chore: move routeParts to global constants.ts file
