+++ src/authentication/views/Logout.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent } from 'react';
 
 import { RouteComponentProps } from 'react-router';
-import { RouteParts } from '../../my-workspace/constants';
+import { RouteParts } from '../../constants';
 import { redirectToLogoutPage } from '../helpers/redirect-to-idp';
 
 export interface LogoutProps {}
