+++ src/constants.ts
@@ -1,0 +1,17 @@
+export const RouteParts = {
+	Search: 'zoeken',
+	Item: 'item',
+	Folder: 'map',
+	Collection: 'collectie',
+	Collections: 'collecties',
+	MyWorkspace: 'mijn-werkruimte',
+	Bookmarks: 'bladwijzers',
+	Folders: 'mappen',
+	Login: 'aanmelden',
+	Logout: 'afmelden',
+	Register: 'registreren',
+	Discover: 'ontdek',
+	Projects: 'projecten',
+	News: 'nieuws',
+	Edit: 'bewerk',
+};
