+++ src/item/routes.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import SecuredRoute from '../authentication/components/SecuredRoute';
-import { RouteParts } from '../my-workspace/constants';
+import { RouteParts } from '../constants';
 import Item from './views/Item';
 
 export const renderItemRoutes = () => (
