+++ src/my-workspace/constants.ts
@@ -1,20 +1,4 @@
-export const RouteParts = {
-	Search: 'zoeken',
-	Item: 'item',
-	Folder: 'map',
-	Collection: 'collectie',
-	Collections: 'collecties',
-	MyWorkspace: 'mijn-werkruimte',
-	Bookmarks: 'bladwijzers',
-	Folders: 'mappen',
-	Login: 'aanmelden',
-	Logout: 'afmelden',
-	Register: 'registreren',
-	Discover: 'ontdek',
-	Projects: 'projecten',
-	News: 'nieuws',
-	Edit: 'bewerk',
-};
+import { RouteParts } from '../constants';
 
 export const COLLECTIONS_ID = RouteParts.Collections;
 export const FOLDERS_ID = RouteParts.Folders;
