+++ src/my-workspace/routes.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 
 import SecuredRoute from '../authentication/components/SecuredRoute';
-import { RouteParts } from './constants';
+import { RouteParts } from '../constants';
 import MyWorkspace from './views/MyWorkspace';
 
 export const renderMyWorkspaceRoutes = () => (
