+++ src/my-workspace/views/MyWorkspace.tsx
@@ -18,8 +18,9 @@ import {
 	ToolbarRight,
 } from '@viaa/avo2-components';
 import Collections from '../../collection/views/Collections';
+import { RouteParts } from '../../constants';
 import ControlledDropdown from '../../shared/components/ControlledDropdown/ControlledDropdown';
-import { BOOKMARKS_ID, COLLECTIONS_ID, FOLDERS_ID, RouteParts, TABS } from '../constants';
+import { BOOKMARKS_ID, COLLECTIONS_ID, FOLDERS_ID, TABS } from '../constants';
 import { MyWorkspaceProps, TabViewMap } from '../types';
 import Bookmarks from './Bookmarks';
