+++ src/search/routes.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import SecuredRoute from '../authentication/components/SecuredRoute';
-import { RouteParts } from '../my-workspace/constants';
+import { RouteParts } from '../constants';
 import Search from './views/Search';
 
 export const renderSearchRoutes = () => (
